requirements.txt
google-api-python-client==2.151.0
youtube-transcript-api==1.2.4
openai==1.60.0
pydantic==2.9.0
pyyaml==6.0.2
python-dotenv==1.0.1
yt-dlp==2025.1.15
deepgram-sdk==3.7.7
pytest==8.3.3
pytest-mock==3.14.0
